add go.mod and dockerfile
