+++ notifier/wechat.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"prometheus-to-webhook/model"
-	"prometheus-to-webhook/transformer"
+	"github.com/k8stech/alertmanager-wechatrobot-webhook/model"
+	"github.com/k8stech/alertmanager-wechatrobot-webhook/transformer"
 )
 
 // Send send markdown message to dingtalk
